[FE] rename: Table 관련 컴포넌트 폴더 이동
why: Table에 필요한 컴포넌트에 응집성을 부여하기 위하여, Table 관련
컴포넌트를 Table 폴더 아래로 옮겼다.

associate with #12
