+++ web/src/component/Table/AddingCard/AddingCard.scss
