+++ web/src/component/Table/Card/card.scss
