+++ web/src/stories/Card.stories.ts
@@ -1,5 +1,5 @@
 import { storiesOf } from "@storybook/html";
-import { Card } from "../component/Card";
+import { Card } from "../component/Table/Card";
 
 storiesOf("Card Storybook", module).add("기본 카드", () =>
   new Card("typescript 공부하기").render()
